track terminal coordinates as (int) rather than (unsigned) in core code.
Use of unsigned terminal coordinates causes problems for clipping.
See for example Bug #2035 where passing unsigned coordinates to do_arc()
caused incorrect treatment of circles with an off-screen center.
This commit changes all calculations of terminal coordinates in the core
code to use signed integers.  It does not change the API for the terminal
driver entry points (term_api.h), which still declares unsigned coordinates.
It also does not change the essentially driver-level code in bitmap.c and
gplt_x11.c.

+++ src/command.c
@@ -1046,10 +1046,10 @@ clear_command()
     term_start_plot();
 
     if (multiplot && term->fillbox) {
-	unsigned int xx1 = (unsigned int) (xoffset * term->xmax);
-	unsigned int yy1 = (unsigned int) (yoffset * term->ymax);
-	unsigned int width = (unsigned int) (xsize * term->xmax);
-	unsigned int height = (unsigned int) (ysize * term->ymax);
+	int xx1 = xoffset * term->xmax;
+	int yy1 = yoffset * term->ymax;
+	unsigned int width = xsize * term->xmax;
+	unsigned int height = ysize * term->ymax;
 	(*term->fillbox) (0, xx1, yy1, width, height);
     }
     term_end_plot();
